Move to esm, tap with parallelized tests, volta for pinning node, and export maps fpr node 14 and above
