+++ .gitignore
@@ -2,3 +2,5 @@ node_modules
 .DS_Store
 *.lock
 *.log
+.nyc_output
+coverage
