+++ test/ansi.mjs
@@ -1,0 +1,1 @@
+export const ANSI = x => `\x1b[${x}m`;
